Restructure `helpers/` directory
The name `helpers/` is consistent with our backend `helpers/` folder. Its just a naming convention really. `utilities/` was not used in the backend system.
