+++ src/components/composites/TextInput/TextInput.tsx
@@ -1,4 +1,4 @@
-import {combineClassNames} from '@utilities/combine-class-names'
+import {combineClassNames} from '@helpers/combine-class-names'
 import React, {forwardRef, useId, JSX} from 'react'
 import styles from './TextInput.module.scss'
 import type {TextNode} from '@nebuloid-types/text-node'
